requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.3.6
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.3.2
cryptography==45.0.6
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2.4
django-filter==25.1
django-import-export==4.3.1
django-rest-passwordreset==1.5.0
django-unfold==0.52.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
factory-boy==3.3.0
Faker==37.5.3
gunicorn==20.1.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
pillow==10.4.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rpds-py==0.26.0
ruff==0.12.5
sentry-sdk==1.45.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.4.0
social-auth-core==4.7.0
sqlparse==0.5.3
tablib==3.7.0
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.9.0
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
